feat: UnitTest를 위한 확장을 Int타입 에서 Double타입으로 변경

+++ Calculator/Calculator/Models/ChooseCalculateItem.swift
@@ -6,4 +6,4 @@
 //
 
 
-extension Int: CalculateItem { }
+extension Double: CalculateItem { }
